+++ packages/backends/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @vercel/backends
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`a9340f6f5022eed96d1e42344b5bb67f1239ca41`](https://github.com/vercel/vercel/commit/a9340f6f5022eed96d1e42344b5bb67f1239ca41)]:
+  - @vercel/introspection@0.0.3
+
+## 0.0.10
+
+### Patch Changes
+
+- Add bun detection for experimental backends ([#14311](https://github.com/vercel/vercel/pull/14311))
+
+- Updated dependencies [[`10de9d8d6ebdef57f4b0ca7dabc277651b317d7c`](https://github.com/vercel/vercel/commit/10de9d8d6ebdef57f4b0ca7dabc277651b317d7c)]:
+  - @vercel/cervel@0.0.6
+  - @vercel/introspection@0.0.2
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [[`06a9b9a8281fa327e9f04a689d512f09eb85201d`](https://github.com/vercel/vercel/commit/06a9b9a8281fa327e9f04a689d512f09eb85201d)]:
+  - @vercel/cervel@0.0.5
+  - @vercel/introspection@0.0.2
+
+## 0.0.8
+
+### Patch Changes
+
+- Fix build issue with experimental backends builder ([#14281](https://github.com/vercel/vercel/pull/14281))
+
+- Updated dependencies [[`cf1183cb80e407238af2dd2a8d5e605f621c5d91`](https://github.com/vercel/vercel/commit/cf1183cb80e407238af2dd2a8d5e605f621c5d91)]:
+  - @vercel/cervel@0.0.4
+  - @vercel/introspection@0.0.2
+
+## 0.0.7
+
+### Patch Changes
+
+- Revert "Remove getSpawnOptions" ([#14261](https://github.com/vercel/vercel/pull/14261))
+
+## 0.0.6
+
+### Patch Changes
+
+- Pull introspection module out of backends for external use ([#14210](https://github.com/vercel/vercel/pull/14210))
+
+- Updated dependencies [[`62651e302de8c2913b658b5cfd5c9e1dd192a03c`](https://github.com/vercel/vercel/commit/62651e302de8c2913b658b5cfd5c9e1dd192a03c)]:
+  - @vercel/introspection@0.0.2
+
 ## 0.0.5
 
 ### Patch Changes